moved impl of try_from for LedgerDeviceType from types.rs to ledger_nano.rs to avoid problems with feature set
